Improve deployment routine wrt pipeline promotions
I've updated the deployment function to take account of the build pack
use of collectstatic, and the DISABLE_COLLECTSTATIC config var. By
default now, if the deployment is a git push, and DISABLE_COLLECTSTATIC
is **not** set, then collectstatic will run by default as part of the
deployment. If it's a pipeline deployment, then the buildpack is
bypassed, and this is not relevant. In this case, if you need to force
collectstatic to run then you must pass the --collectstatic option when
calling deploy:

    heroku-tools deploy app_name --collecstatic

As an example - this is a standard deployment without DISABLE_COLLECSTATIC
set:

  Git branch:    dev
  Target env:    app_name
  Force push:    False
  Run buildpack: True
  Collectstatic: True

And with DISABLE_COLLECTSTATIC set:

  Git branch:    dev
  Target env:    app_name
  Force push:    False
  Run buildpack: True
  Collectstatic: False

And a pipeline deployment:

  Git branch:    dev
  Target env:    app_name
  Force push:    False
  Pipeline:	   True
  Promote: 	   upstream_app_name

And with --collectstatic invoked:

  Git branch:    dev
  Target env:    app_name
  Force push:    False
  Pipeline:	   True
  Promote: 	   upstream_app_name

  ----- Post-deployment commands ------

  collectstatic: python manage.py collectstatic --noinput