Remove automatic migration detection and execution; support migrations via configuration instead
This changeset removes all code which previously autodetected the need to run migrations and then call `manage.py migrate` on the target app.  This behaviour did not play well with pipelined deployments that features multiple apps in a node.

Instead, post-deployment tasks (migrations being the key example) must now be specified via configuration, as a list of shell commands to run via a `post_deploy` node in the config YAML file.

See `settings/sample.conf` for an example of this.

Other changes this squashed commit

* version bump to 0.3 - this is a breaking change
* add a changelog file
* update gitignore to exclude build, dist and coverage dirs